Added a Posts resource controller and moved around other controllers
