+++ app/Http/Controllers/Fos/CategoriesController.php
