+++ app/Http/Controllers/Fos/PostController.php
@@ -1,20 +1,23 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Fos;
 
+use App\Http\Controllers\Controller;
 use App\Models\Post;
 use Illuminate\Http\Request;
 
-class PostsController extends Controller
+class PostController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\View
      */
     public function index()
     {
-        //
+        $posts = Post::all()->sortBy('is_published');
+
+        return view('fos.posts.index', ['posts' => $posts]);
     }
 
     /**
@@ -42,11 +45,11 @@ class PostsController extends Controller
      * Display the specified resource.
      *
      * @param  Post  $post
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\View\View
      */
     public function show(Post $post)
     {
-        //
+        return view('fos.posts.show', ['post' => $post]);
     }
 
     /**
